+++ block-node/spi/src/main/java/module-info.java
@@ -15,7 +15,6 @@ module org.hiero.block.node.spi {
     requires transitive com.swirlds.metrics.api;
     requires transitive org.hiero.block.protobuf;
     requires transitive io.helidon.webserver;
-    requires org.hiero.block.common;
     requires com.github.luben.zstd_jni;
     requires static transitive com.github.spotbugs.annotations;
     requires static java.compiler; // javax.annotation.processing.Generated
