+++ block-node/spi/src/main/java/org/hiero/block/node/spi/threading/ThreadPoolManager.java
@@ -2,10 +2,8 @@
 package org.hiero.block.node.spi.threading;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.concurrent.BlockingQueue;
+import edu.umd.cs.findbugs.annotations.Nullable;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.TimeUnit;
 
 /**
  * An interface that defines a manager for creating and managing thread pools
@@ -18,27 +16,24 @@ import java.util.concurrent.TimeUnit;
  */
 public interface ThreadPoolManager {
     /**
-     * Factory method. Creates a new {@link ExecutorService}.
+     * Factory method. Creates a new single thread {@link ExecutorService}.
      *
-     * @param corePoolSize the core pool size of the thread pool, must be
-     * greater than 0 and less than or equal to maxPoolSize
-     * @param maxPoolSize the maximum pool size of the thread pool, must be
-     * greater 0 and greater than or equal to corePoolSize
-     * @param keepAliveTime the time to keep the threads alive when idle, must
-     * be positive or 0 (zero)
-     * @param timeUnit the time unit of the keep alive time, must not be null
-     * @param workQueue the work queue to use for the thread pool, must not be
-     * null
-     * @param threadFactory the thread factory to use for the thread pool, must
-     * not be null
-     * @return a new instance of the BlockNodeContext record
+     * @param threadNamePrefix the prefix, must not be blank
+     * @return a new single thread executor service
      */
     @NonNull
-    ExecutorService createExecutorService(
-            int corePoolSize,
-            int maxPoolSize,
-            long keepAliveTime,
-            @NonNull final TimeUnit timeUnit,
-            @NonNull final BlockingQueue<Runnable> workQueue,
-            @NonNull final ThreadFactory threadFactory);
+    ExecutorService createSingleThreadExecutor(@NonNull final String threadNamePrefix);
+
+    /**
+     * Factory method. Creates a new single thread {@link ExecutorService} using
+     * the specified (nullable) {@link Thread.UncaughtExceptionHandler}.
+     *
+     * @param threadNamePrefix the prefix, must not be blank
+     * @param uncaughtExceptionHandler the uncaught exception handler, nullable
+     * @return a new single thread executor service
+     */
+    @NonNull
+    ExecutorService createSingleThreadExecutor(
+            @NonNull final String threadNamePrefix,
+            @Nullable final Thread.UncaughtExceptionHandler uncaughtExceptionHandler);
 }
